ecoxplorer-api/requirements.txt
openai==1.51.2
boto3==1.35.36
python-dotenv==1.0.1